Fixed a couple of memory leaks.

+++ sequential/src/steiner.c
@@ -10,8 +10,6 @@
 #include "common.h"
 #include "shortestpath.h"
 
-clock_t elaps;
-
 int steiner_verification(graph* g, int v, void* input) {
 
     int check1 = degree(g, v) >= 3;
@@ -27,7 +25,7 @@ pair* steiner_tree(graph* g, set_t* terminals) {
     int T = set_size(terminals);
     long long P =  (long long) pow(2, T) - 1;
 
-    int** costs = (int**) malloc(sizeof(int*) * V);
+    int** costs  =  (int**) malloc(sizeof(int*) * V);
     graph*** trees = (graph***) malloc(sizeof(graph**) * V);
     
     for(int v=0; v < V; v++) {
@@ -39,37 +37,16 @@ pair* steiner_tree(graph* g, set_t* terminals) {
 
     // All pairs shortest path
 
-    elaps = clock();
     pair* apsp = all_pairs_shortest_path(g);
-    printf("All Pairs Shortest Path: %fs.\n", (double) (clock() - elaps)/CLOCKS_PER_SEC);
 
     graph*** paths  = (graph***) apsp->first;
     int** distances =    (int**) apsp->second;
 
     free(apsp);
 
-    // Fill base cases
+    // Fill tables
 
-    for(int v=0; v < V; v++) {
-
-        long long mask = 1ll << (T - 1);
-
-        for(int pos=0; pos < T; pos++) {
-
-            int u = g->reverse_hash[get_element(terminals, pos)];
-
-            costs[v][mask - 1] = distances[v][u];
-            trees[v][mask - 1] = paths[v][u];
-
-            mask = mask >> 1;
-
-        }
-
-    }
-
-    elaps = clock();
-
-    for(int k=2; k <= T; k++) {
+    for(int k=1; k <= T; k++) {
 
         long long mask = 0;
 
@@ -79,41 +56,55 @@ pair* steiner_tree(graph* g, set_t* terminals) {
 
             for(int v=0; v < V; v++) { // O(T * 2^T * V * (V + E))
 
-                set_t* W = bfs(g, g->hash[v], steiner_verification, X); // O(V+E)
+                if(k == 1) {
+
+                    int u = g->reverse_hash[get_element(X, 0)];
+
+                    costs[v][mask - 1] = distances[v][u];
+                    trees[v][mask - 1] = paths[v][u];
+
+                } else {
+
+                    set_t* W = bfs(g, g->hash[v], steiner_verification, X); // O(V+E)
             
-                int min = INT_MAX;
-                graph* min_tree = NULL;
+                    int min = INT_MAX;
+                    graph* min_tree = NULL;
 
-                for(int i=0; i < set_size(W); i++) { // O(T * 2^T * (V+E))
+                    for(int i=0; i < set_size(W); i++) { // O(T * 2^T * (V+E))
 
-                    int w = g->reverse_hash[get_element(W, i)];
+                        int w = g->reverse_hash[get_element(W, i)];
 
-                    if( element_exists(g->hash[w], X) ) { // O(V + E)
+                        if( element_exists(g->hash[w], X) ) { // O(V + E)
 
-                        long long submask = 1ll << (T - find_position(terminals, g->hash[w]) - 1);
+                            long long submask = 1ll << (T - find_position(terminals, g->hash[w]) - 1);
 
-                        int cost = distances[v][w] + costs[w][(mask & ~submask) - 1];
-
-                        if(cost < min) {
-                            min = cost;
-                            min_tree = graph_union(paths[v][w], trees[w][(mask & ~submask) - 1]); // (V + E)
-                        }
-
-                    } else { // O(2^T (V+E))
-
-                        for(long long submask = (mask - 1) & mask; submask != 0; submask = (submask - 1) & mask) { // iterate over submasks of the mask O(2^T)
-
-                            int cost = distances[v][w] + costs[w][submask - 1] + costs[w][(mask & ~submask) - 1];
+                            int cost = distances[v][w] + costs[w][(mask & ~submask) - 1];
 
                             if(cost < min) {
-
                                 min = cost;
 
                                 destroy_graph(min_tree);
+                                min_tree = graph_union(paths[v][w], trees[w][(mask & ~submask) - 1]); // (V + E)
 
-                                graph* tmp_tree = graph_union(trees[w][submask - 1], trees[w][(mask & ~submask) - 1]); // O(V + E)
-                                min_tree = graph_union(paths[v][w], tmp_tree);
-                                destroy_graph(tmp_tree);
+                            }
+
+                        } else { // O(2^T (V+E))
+
+                            for(long long submask = (mask - 1) & mask; submask != 0; submask = (submask - 1) & mask) { // iterate over submasks of the mask O(2^T)
+
+                                int cost = distances[v][w] + costs[w][submask - 1] + costs[w][(mask & ~submask) - 1];
+
+                                if(cost < min) {
+
+                                    min = cost;
+
+                                    destroy_graph(min_tree);
+
+                                    graph* tmp_tree = graph_union(trees[w][submask - 1], trees[w][(mask & ~submask) - 1]); // O(V + E)
+                                    min_tree = graph_union(paths[v][w], tmp_tree);
+                                    destroy_graph(tmp_tree);
+
+                                }
 
                             }
 
@@ -121,13 +112,13 @@ pair* steiner_tree(graph* g, set_t* terminals) {
 
                     }
 
+                    costs[v][mask - 1] = min;
+                    trees[v][mask - 1] = min_tree;
+
+                    destroy_set(W);
+
                 }
-
-                costs[v][mask - 1] = min;
-                trees[v][mask - 1] = min_tree;
-
-                destroy_set(W);
-
+                
             }
 
             destroy_set(X);
@@ -135,10 +126,6 @@ pair* steiner_tree(graph* g, set_t* terminals) {
         }
 
     }
-
-    printf("DP step: %fs.\n", (double) (clock() - elaps)/CLOCKS_PER_SEC);
-
-    // print_table(costs, V, P);
 
     // Extract minimum from table.
 
@@ -160,12 +147,31 @@ pair* steiner_tree(graph* g, set_t* terminals) {
     free_table(costs, V, P);
     free_table(distances, V, V);
 
+    for(int i=0; i<V; i++) {
+
+        for(int j=0; j<V; j++) {
+            destroy_graph(paths[i][j]);
+            paths[i][j] = NULL;
+        }
+
+        free(paths[i]);
+        paths[i] = NULL;
+
+    }
+
+    free(paths);
+    paths = NULL;
+
     for(int i=0; i < V; i++) {
 
         for(long long j=0; j<P; j++) {
 
-            if(i != min_index || j != P - 1) {
-                destroy_graph(trees[i][j]);
+            if(__builtin_popcount(j + 1) != 1) {
+
+                if(i != min_index || j != P - 1) {
+                    destroy_graph(trees[i][j]);
+                }
+
             }
 
             trees[i][j] = NULL;
